Make main plugin file PHP 5.6 compatible and move main plugin functionality in a separate file
